Ensure that monitor always exists
Some methods, such as #touch, don't trigger the `before_update`
callback. Because of that, `@nested_attribute_destruction_monitor` was
nil if such methods were called before any method that does trigger the
callback.

Setting up the monitor after initialization makes sure that it is always
available if we need it.